Change amout type in transaction

+++ app/models/transaction.rb
@@ -3,6 +3,6 @@ class Transaction
   include Mongoid::Timestamps
   field :user_id, type: Integer
   field :name, type: String
-  field :amount, type: Integer
+  field :amount, type: BigDecimal
   field :kind, type: String
 end
